user-booking/src/main/java/com/user/booking/service/UserServiceImpl.java
package com.user.booking.service;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

import com.user.booking.model.User;
import com.user.booking.repository.UserRepository;

/**
 * 
 * Implementation class for User Service
 *
 */
@Service
public class UserServiceImpl implements UserService {

	private static final Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);

	@Autowired
	private UserRepository userRepository;

	@Autowired
	private BCryptPasswordEncoder bCryptPasswordEncoder;

	@Override
	public void save(User user) {
		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
		logger.info("Roless  " + user.getRoles());
		user.setRoles(user.getRoles());

		userRepository.save(user);
	}

	@Override
	public User findByUsername(String username) {
		return userRepository.findByUsername(username);
	}

	@Override
	public Boolean existsByUsername(String username) {
	
		return null;
	}

	@Override
	public Boolean existsByEmail(String email) {
	
		return null;
	}

}
